tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.attributes.shadowColor.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.attributes.shadowColor.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.attributes.shadowColor.valid</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.shadowColor = 'lime';
  _assertSame(ctx.shadowColor, '#00ff00', "ctx.shadowColor", "'#00ff00'");

  ctx.shadowColor = 'RGBA(0,255, 0,0)';
  _assertSame(ctx.shadowColor, 'rgba(0, 255, 0, 0)', "ctx.shadowColor", "'rgba(0, 255, 0, 0)'");
  t.done();

});
</script>
